# backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductReadDto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductReadDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ProductDetail/{alias}": {
      "get": {
        "tags": [
          "Products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "alias",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductReadDto"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
...
    "OrderItem": {
      "required": [
        "orderID",
        "productSizeID",
        "quantity"
      ],
      "type": "object",
      "properties": {
        "orderItemID": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "orderID": {
          "format": "int32",
          "type": "integer"
        },
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Product": {
      "required": [
        "categoryID",
        "name"
      ],
      "type": "object",
      "properties": {
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "brand": {
          "maxLength": 50,
          "type": "string"
        },
        "description": {
          "maxLength": 500,
          "type": "string"
        },
        "imageURL": {
          "format": "uri",
          "type": "string"
        },
        "nameAlias": {
          "type": "string"
        },
        "categoryID": {
          "format": "int32",
          "type": "integer"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "productSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductSize"
          }
        }
      }
    },
    "ProductCreateDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "categoryID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ProductReadDto": {
      "type": "object",
      "properties": {
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "nameAlias": {
          "type": "string"
        },
        "categoryID": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "productSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductSizeReadDto"
          }
        }
      }
    },
    "ProductSize": {
      "required": [
        "price",
        "productID",
        "quantity",
        "size"
      ],
      "type": "object",
      "properties": {
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        }
      }
    },
...
    "ProductSizeReadDto": {
      "type": "object",
      "properties": {
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        }
      }
    },
...
    "ProductUpdateDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "categoryID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}